requirements.txt
numpy
opencv-python
torch
paho-mqtt
pyyaml
